Update web worker for chat
