--- src/api/nitro/AddWorkerEventTracker.ts
@@ -1,7 +1,0 @@
-import { IWorkerEventTracker } from '@nitrots/nitro-renderer';
-import { GetNitroInstance } from './GetNitroInstance';
-
-export const AddWorkerEventTracker = (tracker: IWorkerEventTracker) =>
-{
-    GetNitroInstance().addWorkerEventTracker(tracker);
-}
