--- src/api/nitro/RemoveWorkerEventTracker.ts
@@ -1,7 +1,0 @@
-import { IWorkerEventTracker } from '@nitrots/nitro-renderer';
-import { GetNitroInstance } from './GetNitroInstance';
-
-export const RemoveWorkerEventTracker = (tracker: IWorkerEventTracker) =>
-{
-    GetNitroInstance().removeWorkerEventTracker(tracker);
-}
