--- src/api/nitro/SendWorkerEvent.ts
@@ -1,6 +1,0 @@
-import { GetNitroInstance } from './GetNitroInstance';
-
-export const SendWorkerEvent = (message: { [index: string]: any }) =>
-{
-    GetNitroInstance().sendWorkerEvent(message);
-}
