+++ src/api/nitro/index.ts
@@ -1,5 +1,4 @@
 export * from './AddLinkEventTracker';
-export * from './AddWorkerEventTracker';
 export * from './avatar';
 export * from './camera';
 export * from './core';
@@ -11,8 +10,6 @@ export * from './GetLocalization';
 export * from './GetNitroInstance';
 export * from './OpenUrl';
 export * from './RemoveLinkEventTracker';
-export * from './RemoveWorkerEventTracker';
 export * from './room';
 export * from './SendMessageComposer';
-export * from './SendWorkerEvent';
 export * from './session';
