+++ src/workers/IntervalWebWorker.ts
@@ -1,9 +1,6 @@
 export default () =>
 {
-    const intervals: {
-        id: number,
-        interval: ReturnType<typeof setInterval>
-    }[] = [];
+    let interval: ReturnType<typeof setInterval> = null;
 
     // eslint-disable-next-line no-restricted-globals
     self.onmessage = (message: MessageEvent) =>
@@ -11,38 +8,19 @@ export default () =>
         if(!message) return;
 
         const data: { [index: string]: any } = message.data;
-    
-        switch(data.type)
+
+        switch(data.action)
         {
-            case 'CREATE_INTERVAL': {
-                const id = (data.timerId as number);
-                const time = (data.time as number);
-                const response = (data.response as string);
-    
-                const interval = setInterval(() => postMessage(response), time);
-    
-                intervals.push({ id, interval });
-                return;
-            }
-            case 'REMOVE_INTERVAL': {
-                const id = (data.timerId as number);
-    
-                const i = 0;
-    
-                while(i < intervals.length)
+            case 'START':
+                interval = setInterval(() => postMessage(null), data.content);
+                break;
+            case 'STOP':
+                if(interval)
                 {
-                    const interval = intervals[i];
-    
-                    if(interval.id === id)
-                    {
-                        clearInterval(interval.interval);
-    
-                        intervals.splice(i, 1);
-    
-                        return;
-                    }
+                    clearInterval(interval);
+                    interval = null;
                 }
-            }
+                break;
         }
     }
 }
